Add comments popup

+++ client/src/components/CommentsBlock/Comment/Comment.tsx
@@ -1,4 +1,4 @@
-import { FC } from "react";
+import { FC, useEffect, useRef, useState } from "react";
 import { IComments } from "../../../types/IComments";
 import classes from "./Comment.module.css";
 import deleteIcon from "../../../assets/img/delete.svg";
@@ -7,6 +7,9 @@ import { RootState, useAppDispatch } from "../../../redux/store";
 import { deleteCommentById } from "../../../redux/slices/commentsSlice";
 import { toast } from "react-toastify";
 import { Link } from "react-router-dom";
+import dots from "../../../assets/img/dots.svg";
+import danger from "../../../assets/img/danger.svg";
+import edit from "../../../assets/img/edit.svg";
 
 type CommentProps = {
   comment: IComments;
@@ -16,12 +19,30 @@ const Comment: FC<CommentProps> = ({ comment }) => {
   const dispatch = useAppDispatch();
   const { user } = useSelector((state: RootState) => state.user);
 
+  const [isOpenActions, setIsOpenActions] = useState(false);
+  const menuRef = useRef(null);
+
+  useEffect(() => {
+    const handleDocumentClick = (event) => {
+      if (isOpenActions && menuRef.current && !menuRef.current.contains(event.target)) {
+        setIsOpenActions(false);
+      }
+    };
+
+    document.addEventListener("click", handleDocumentClick);
+
+    return () => {
+      document.removeEventListener("click", handleDocumentClick);
+    };
+  }, [isOpenActions]);
+
   const deleteComment = (id: number) => {
-    toast.promise(dispatch(deleteCommentById(id)), {
-      pending: "Загрузка...",
-      success: "Комментарий успешно удален",
-      error: "Ошибка удаления комментария",
-    });
+    confirm("Вы действительно хотите удалить комментарий?") &&
+      toast.promise(dispatch(deleteCommentById(id)), {
+        pending: "Загрузка...",
+        success: "Комментарий успешно удален",
+        error: "Ошибка удаления комментария",
+      });
   };
 
   return (
@@ -37,13 +58,37 @@ const Comment: FC<CommentProps> = ({ comment }) => {
             <Link to={`/profile/${comment.user_id}`}>{comment.name}</Link>
             <p className={classes.date}>{comment.created_at}</p>
           </div>
-          <div className={classes.dots}>...</div>
+          <div
+            className={classes.dots}
+            onClick={(e) => {
+              e.stopPropagation();
+              setIsOpenActions(!isOpenActions);
+            }}>
+            <img src={dots} alt="dots" />
+          </div>
         </div>
       </div>
       <div className={classes.commentText}>{comment.text}</div>
-      {(+user.id === comment.user_id || user.role === "admin") && (
-        <div className={classes.deleteIcon} onClick={() => deleteComment(comment.id)}>
-          <img src={deleteIcon} alt="delete" />
+      {isOpenActions && (
+        <div className={classes.commentActions} ref={menuRef}>
+          <ul className={classes.actions}>
+            <li className={classes.actionItem}>
+              <img src={danger} className={classes.deleteIcon} />
+              Пожаловаться
+            </li>
+            {(+user.id === comment.user_id || user.role === "admin") && (
+              <>
+                <li className={classes.actionItem}>
+                  <img src={edit} className={classes.deleteIcon} />
+                  Редактировать
+                </li>
+                <li className={classes.remove} onClick={() => deleteComment(comment.id)}>
+                  <img src={deleteIcon} className={classes.deleteIcon} />
+                  Удалить
+                </li>
+              </>
+            )}
+          </ul>
         </div>
       )}
     </div>
